[DLTv2][BUG] Support variable-length ID and fix memory issues
This commit addresses multiple critical issues in the DLT Protocol V2
implementation to properly support AUTOSAR spec variable-length IDs
(1-255 bytes with length prefix, no null terminator):

Memory Safety Fixes:
- Fix heap-buffer-overflow in dlt_message_read_v2 when receiving messages
  with variable-length IDs
- Fix stack-use-after-scope in dlt_client_send_ctrl_msg_v2 by moving
  ID buffers to function scope
- Fix uninitialized DltMessageV2 structure causing crashes by properly
  initializing extendedheaderv2 to zero
- Fix memory leaks: free allocated tags in dlt_message_free_v2 and
  dlt_message_read_v2, free headerbuffer when reusing message objects

Protocol Implementation:
- Add storage header V2 parsing in dlt_message_read_v2 to handle FIFO
  messages (pattern "DLT\x02")
- Remove illegal NULL terminators from wire format - IDs are length-prefixed
  binary data, not C strings
- Fix daemon internal logging to use dynamic ECU ID length (strlen) instead
  of hardcoded DLT_V2_ID_SIZE (255)
- Fix dangling pointers by pointing to global data instead of temporary
  stack buffers

Display and Forwarding:
- Implement variable-length ID display up to DLT_CLIENT_MAX_ID_LENGTH
  (16 bytes) for ECU/APP/CTX IDs
- Fix spacing in header display (single space between fields)
- Fix daemon forwarding by updating baseheaderv2 pointer after buffer
  reallocation

Testing:
- Fix test to call dlt_file_free_v2 instead of dlt_file_free for v2 files

+++ src/console/dlt-receive-v2.c
@@ -662,13 +662,13 @@ int dlt_receive_message_callback_v2(DltMessageV2 *message, void *data)
     else
         dlt_set_storageheader_v2(&(message->storageheaderv2), dltdata->ecuidlen, dltdata->ecuid);
 
-    message->storageheadersizev2 = (uint8_t)(STORAGE_HEADER_V2_FIXED_SIZE + message->storageheaderv2.ecidlen);
+    message->storageheadersizev2 = (uint32_t)(STORAGE_HEADER_V2_FIXED_SIZE + message->storageheaderv2.ecidlen);
 
     /* Add Storage Header to Header Buffer and update header size*/
     uint8_t temp_buffer[message->headersizev2];
     memcpy(temp_buffer, message->headerbufferv2, (size_t)message->headersizev2);
     free(message->headerbufferv2);
-    message->headersizev2 = (uint8_t)(message->headersizev2 + (int32_t)message->storageheadersizev2);
+    message->headersizev2 = message->headersizev2 + (int32_t)message->storageheadersizev2;
 
     message->headerbufferv2 = (uint8_t *)malloc((size_t)message->headersizev2);
